Adapt the calculator to work in CLI mode
When called without arguments, keep the same behaviour as before. If
arguments are passed, interpret them as a sequence of commands, perform
the operations and print out the final state of the stack.

Some examples:
    $ python rc.py 3 3 +
    6.0
    $

    $ python rc.py 3 2 5 +
    3.0
    7.0
    $

    $ python rc.py 3 drop
    $

    $ python rc.py 3 +
    ***
    Not enough arguments for operator '+'
    ***

In this mode, errors are output to stderr, and nothing is written to
stdout to avoid incoherent state of the stack.